requirements.txt
fastapi
uvicorn
python-dotenv
plivo
python-multipart
jinja2